cosmetic renaming of rethinkdriver -> rethinkDriver
