+++ src/server/graphql/models/CachedUser/helpers.js
@@ -1,4 +1,4 @@
-import r from '../../../database/rethinkdriver';
+import r from '../../../database/rethinkDriver';
 
 export const getUserByUserId = async userId => {
   const users = await r.table('CachedUser').getAll(userId, {index: 'userId'}).limit(1);
