+++ src/server/graphql/models/Meeting/helpers.js
@@ -1,3 +1,3 @@
-import r from '../../../database/rethinkdriver';
+import r from '../../../database/rethinkDriver';
 
 export const getMeetingById = async id => await r.table('Meeting').get(id);
