+++ src/server/graphql/models/Team/helpers.js
@@ -1,3 +1,3 @@
-import r from '../../../database/rethinkdriver';
+import r from '../../../database/rethinkDriver';
 
 export const getTeamById = async id => await r.table('Team').get(id);
