+++ src/server/graphql/models/Team/teamMutation.js
@@ -1,5 +1,5 @@
 import { Team } from './teamSchema';
-import r from '../../../database/rethinkdriver';
+import r from '../../../database/rethinkDriver';
 import uuid from 'node-uuid';
 import {
   GraphQLString,
