+++ src/server/graphql/models/TeamMember/teamMemberMutation.js
@@ -1,5 +1,5 @@
 import { TeamMember } from './teamMemberSchema';
-import r from '../../../database/rethinkdriver';
+import r from '../../../database/rethinkDriver';
 import uuid from 'node-uuid';
 import {
   GraphQLString,
